backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client>=1.7.0
openai==1.3.4
psycopg2-binary
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
slowapi==0.1.8